Split GpApiConnector into categories
