+++ GlobalPayments-iOS-SDK/Classes/Gateways/GlobalPayments/GpApiConnector+SerializeRequest.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+extension GpApiConnector {
+
+    func serializeRequest(_ builder: AuthorizationBuilder) -> String? {
+        print("Not implemented yet")
+        abort()
+    }
+}
